pip_requirements.txt
harmony-service-lib~=1.0.22
pystac~=0.5.6
matplotlib==3.7.1
rasterio==1.3.6
numpy==1.24.2
